Switch score from float to int16_t

+++ include/photon/engine/eval.h
@@ -12,7 +12,8 @@ namespace photon::engine {
 
 struct evaluation_t {
 	result_t result;
-	float score;
+	/** Score in centipawns */
+	int16_t score;
 	/** Principal variation in reverse order */
 	std::vector<move_t> moves;
 };
@@ -36,7 +37,7 @@ using evalstate_ptr_t = std::unique_ptr<evalstate_t, evalstate_deleter_t>;
 
 evalstate_ptr_t CreateEvalState();
 
-float PositionHeuristic(const board_t& board);
+int16_t PositionHeuristic(const board_t& board);
 
 std::pair<evaluation_t, evalmetrics_t> EvalBoard(const board_t& board, const searchparams_t& params);
